Fix: passive AD_MAGM damage when attacking the Oracle (player or another monster).
This is a long-standing bug, especially concerning pets or other
monsters attacking the Oracle. The order of events were wrong - the
attacking pet/monster would take passive AD_MAGM (magic missile) first,
and then the actual attack would occur. This was because 'mhit' was not
defined for that type of passive attack. Oops. It was really that
simple. While there, I noticed that monsters that fall under
resists_mgc() were not accounted for, so that'd been included.

While working on this, I noticed that the same conditions for the player
attacking the Oracle could use some attention also.