Fix: Giantslayer strength bonus not taken into account whilst twoweaponing.
The player would still receive 25 strength when wielding Giantslayer in
either the primary or offhand while twoweaponing, but the maximum weight
restriction was not negated like it would be if wearing gauntlets of
power. So, adjust the check away from objects, and have it check the
players actual strength.

Snuck in a grammar fix as well.