Fix: cursed Ogresmasher hitting oneself and knockback.
If wielding a cursed weapon and not an Infidel, there's a 1 in 7 chance
of hitting yourself in the face when attacking with that weapon. The
artifact weapon Ogresmasher has a chance to knockback its target on a
successful hit. A crash would occur if you hit yourself with cursed
Ogresmasher and knockback kicked in (hurtle() doesn't account for you
being the attacker and the target). Simple fix is to just disallow the
knockback feature to fire if Ogresmasher is cursed. A better fix might
be to allow hurtle() to count you as the attacker and target at the same
time, but that's more work than I'm willing to take on right at this
moment.

Also in this commit - Ogresmasher's knockback effect will not work if
you're engulfed (defending), or if the target is unsolid.