Additional feedback when damaging monsters vulnerable to elemental damage.
Provide additional feedback when damaging a monster that is vulnerable
to various types of elemental damage. Example - hitting a wolf with Fire
Brand burns it. Hitting a blue jelly with the same will 'severely burn'
it. This will give the (unspoiled) player some clue as to what elemental
damage types can do additional damage to a monster.