Show in inventory if the Bell of Opening can't be used (quest not complete).
If the player's quest is not complete, the Bell of Opening will not
function. I've seen some players forget this and try to use the Bell
before their quest is finished. This commit adds a bit of an identifier
to the Bell when viewed, it'll be appended with (locked) if it can't be
used. As you'd suspect, (locked) goes away once the quest is finished.

If anyone has a better suggestion for (locked), I'm all ears.