Various monsters that can spawn in groups will sometimes be lead by a stronger leader.
Saw this in practice when I ascended GnollHack last year - groups of
monsters would be lead by a higher level monster of the same type (group
of hill orcs lead by an orc-captain for example). The way this is
achieved in that variant is... not ideal. So I decided to work on it
myself. Had the code written, preliminary testing was looking good, and
then I thought to myself 'didn't SpliceHack already do this?' and it
turns out it did, and was along the same lines of what I had put
together. So I took some of its code, merged it with mine and made some
improvements.

So the gist of it is this: various monsters that can spawn in groups can
have a higher level 'leader' spawn along side, taking up one of the
slots that would have been one of the weaker orignal monsters. The code
uses little_to_big() to determine what the leader monster will be. So a
large group of hill orcs will also include a single orc-captain to lead
them. Or a pack of winter wolf cubs will have a single adult winter wolf
present. I excluded a few animal types due to some of the jumps in
little_to_big() being too great (example: giant rat to enormous rat,
encountering a group of giant rats on dlvl 1 with an enormous rat
leading the pack starting out would be way too difficult).

Along with a leader, some monster groups can also spawn with a 'support'
monster. This support monster will hang back and not engage the
player unless the players hit points drop to a certain level. This
support monster is a caster, and will heal the other monsters in its
group when they need it. I left wizard-type monster casters out of this
commit on purpose, only cleric types are included. Some of these support
casters aren't too much of a threat - group of kobolds, kobold shaman
for instance. Come across a group of drow raiders or gnoll hunters
though, and their support caster would be a cause for concern.

This commit definitely ramps up the difficulty, but will also make for
some very fun, interesting and challenging scenarios.