Fix: more fixes in mcastu.c
There were a few places in the routines that handles you casting a
monster spell or a monster casting vs another monster, where it was only
taking the players level into account for damage/duration. Fixed so it
takes either the players level or the casting monsters level as
necessary.