Pets can help the player other than fighting under certain cirmcumstances.
Saw this type of behavior initially in Splicehack - if you chose the
role of a Dragon Rider, your pet dragon would assist you with various
troubles. Using that as a model, I created my own pet assists for
EvilHack. The breakdown:

* Pets that can cast clerical spells (AD_CLRC) can heal the player if
the players hit points drop to 1/8th or less of their maximum value
* Pets that can cast clerical spells and are of a high enough level will
cast protection on the player if not already protected via the spell
* Pets that can cast wizard spells (AD_SPEL) and are of a high enough
level will cast reflection on the player if they don't already have
reflection
* Pets that can have a breath or gaze fire attack can unslime the
player. The player will take a small amount of fire damage if not fully
resistant
* Pets can free the player from various traps, the critera being that
the pet is strong, has limbs, and for some traps is at least the same
size of the player or larger. The player can take some physical damage
if being ripped free of a bear trap

Not all of this behavior happens immediately every time (protection,
reflection, trap removal), and I may tweak the frequency as needed. And
some of this behavior comes with a certain amount of risk (umslimed,
bear trap removal) to keep things interesting.

I may add more to this, we'll see. I think this is a good start. Will
mainly be keeping an eye on the healing bit, need to make sure that
doesn't become a free ride and the player never having to worry about
their hit points getting to zero.