Tweaks to altar sacrifice gifts.
This is part of a planned overhaul to altar sacrifice gifts, this bit
doesn't break saves. It also fixes a minor issue as well.

First off: Wizards. Because of the limited number of weapons they can
be proficient in, what they could receive as a weapon needed to be
changed, otherwise they could have received a weapon they can't reliably
use. Also, if a Wizard keeps 'never hit with a wielded weapon'
conduct intact, they'll be gifted a spellbook (divine knowledge or the
spellbook, circumstance dependent) instead of a weapon.

If you've never abused your alignment, you have a greater chance of a
regular gift having an object property.

Last bit - there's a school of thought that states that using for
example rn2(2) isn't really as random as doing rn2(100) >= 50. You'll
see this mentioned in comments in the actual code from the NetHack
devteam. So while here, I decided to apply this reasoning to how various
objects here are randomly selected.