Fix: various 'hiding under' bugs.
Started fuzzing again lately, and the issues that have been
consistiently triiping up the fuzzer have been tied to you or a monster
hiding under an object that is no longer there. EvilHack already had a
few fixes for these bugs from NetHack 3.7, but I've lagged behind a bit.
This commit fixes the remaining known bugs in this regard.