Fix: random amulet materials
When looking at objects via blessed object detection, amulets carried by monsters would show as having random materials, changing each time the cursor moved to them. This happened because dummy items would be created so a description could be given, and those items would have a random material according to the usual distribution.

It's still unclear to me why the dummy object needs to be created in this case (compared to when the amulet is just on the ground), but I don't want to mess with it, and I think it makes sense for the dummy objects to always be default material since they're supposed to be as generic as possible, so this change should be done anyway.