Fix: crashes when naming wielded objects in some circumstances
A result of commit 2fba95d. mbodypart() could be called with a null monster, resulting in a crash, if doname() was called on a wielded object that was freed from inventory (most easily replicable by dipping a single wielded potion into another potion). Instead, always assume the player is the wielder unless the object is explicitly held by a monster.

Also, add a check in mbodypart() to prevent crashes if it somehow gets called with a null monster.