Fix: digesting rings of fire/cold/shock/poison resistance would not confer any resistance.
This bug has most likely existed since the very beginning of EvilHack
development when partial resistances were ported over from SporkHack
(and this bug exists there as well). Was noticed recently during a game
of Hack'EM, which shares a lot of EvilHack code. A fix was made for that
variant; the fix here is similar, but not exactly the same.

It was pointed out by terrapin that a more elegant fix would be to
change a bit in how_resistant() in potion.c from
u.uprops[which].intrinsic & (FROMEXPER | FROMRACE | FROMFORM) to
u.uprops[which].intrinsic & INTRINSIC, but doing so would break a lot of
the functionality of ice traps. So we've made specific cases for each of
the rings affected instead. Eating any of these rings has a chance to
grant 50% resistance for that type of ring (one third of that for fire
res if Draugr race).