Fix: 'zaps you with a !'
It's rare, but a shambling horror could spawn with AT_MAGC/AD_ACID
combo, and that's exactly what happened when this minor bug was
discovered. There was a missing entry for that attack type combo and
feedback for the damage type being acid. Easy fix. While there, went
ahead and filled in the missing entries for other damage types as well,
in case those are ever made into that type of attack/damage form also.