backport 3.7 makedefs options for producing individual files
This may be useful for some build environments to avoid parallel make
issues, and artificially-concocted order dependencies, leaving the ordering
up to that specified in the Makefile.

This does not update the 3.6 sys/unix/Makefile.dat, since there's
already a workaround in place in that file.

The related makedefs options are:

       -s     Generate the bogusmon , engrave and epitaph files.

       -1     Generate the epitaph file.

       -2     Generate the engrave file.

       -3     Generate the bogusmon file.