Fix: prevent lawful/neutral players from receiving an altar sacrifice weapon gift with the venom object property.
There was already a check in place to prevent regular altar sacrifice
gifts from being coated in poison for lawful/neutral players, but there
wasn't any kind of check for a weapon having the venom object property.
If that happens, strip that object property and replace with one that
the player can use without abusing their alignment.