New material type - Bronze.
Made this material type on a whim of sorts. A few of us were talking
about how neat it could be if vanilla NetHack had bronze weapons and
armor (aside from bronze plate mail), and then make a conduct based on
only using bronze-age or older weapons and gear. So why not include an
actual separate material 'bronze' to the game? So now, 'bronze plate
mail' can be a thing again.

Weight and AC protection is comparable to iron, cost is slightly more.
Copper had its cost adjusted slightly upward.

A bit of code formatting in here as well.