Finish up Druid quest text, fix formatting.
Made a few small edits from the initial Druid text commit, but this is
mainly fixes to formatting. I also forgot to flag Elanee as a proper
name, so that's been addressed as well.

Maps still need to be completed, but I made some minor edits to the .des
file for testing purposes.