All demons/devils are now unaligned (evil).
A monster having an alignment of unaligned has always denoted being
'evil'. In regular NetHack, this would only be the Wizard of Yendor, and
if memory serves, the high priest in Moloch's temple in the Valley of
the Dead. In EvilHack, it's included a few more monsters, such as the
Archbishop of Moloch and their attendants (cultists), player monster
Infidels, and of course player Infidels, which become demons when
crowned.

Demons and devils in NetHack are either chaotic (demons) or lawful
(almost every devil), to include the unique demon/devil bosses. I
understand why this is the way it is, because in advanced Dungeons &
Dragons, demons are always chaotic evil, and devils lawful evil. But
NetHack doesn't have the nine alignment system, just the three -
lawful/neutral/chaotic. This doesn't really work well for demons and
devils in the world of NetHack, and especially in a variant like
EvilHack where 'true evil' means a lot more than in other variants.
Side note - horned/barbed devils would never randomly spawn in Gehennom,
because they are lawful. Technically, this is a bug. Maybe the NetHack
devs will address that someday.

Anywho. This commit makes all demons and devils unaligned, or 'evil'. As
they should be. It's not as simple as just changing their alignment,
some edits also needed to be made for artifact handling for
demons/devils, and the monster spell 'summon nasties'. Because of this
change, demons can now gate in all types of demons/devils, and are not
restricted by their alignment. Also, horned/barbed devils can now
randomly appear in Gehennom, like they should have been able to do from
the beginning.