Fix: monster stealing worn gold objects from other monsters (AD_SGLD) not properly handled.
This was discovered when a player's steed wearing gold barding had it
stolen by a leprechaun. When trying to fit another set of barding onto
their steed, they couldn't, as the oworkmask hadn't been properly
cleared. This commit fixes that. To be clear, this isn't just for
barding - this would apply to any piece of gear worn or weapon wielded
that was made of gold that was stolen.