New object - helm of caution, change 'shatterproof' to 'tempered'.
This was initially just going to be changing 'shatterproof' to
'tempered' for fixed glass objects, as this is how it's called in
NetHack 3.7 (devteam uses 'cracked' instead of 'fractured' for damaged
glass objects, but oh well). That's when I noticed the new helm added to
3.7, the helm of caution, which grants warning. And the helm of
brilliance was changed to be made of glass. So I figured why not add all
that too? So here we are.