Change the recipe for the forged artifact Ithilmar.
The old recipe did two things: it made forging Ithilmar very difficult
to do in terms of all of the artifacts needed to create it, and one of
the artifacts needed to make it (Gauntlets of Purity) was rather
difficult to give up. Swapped out the gauntlets with Dramborleg, which
hasn't been used in any previous recipies, and its magic resistance
fits with the abilities Ithilmar can impart to the steed wearing it.