Fix: being engulfed using lit/unlit state of tile the player was standing on when engulfed.
The issue - player is engulfed by a monster, they are now inside its
stomach or what have you. Player (correctly) assumes that they are now
in the dark. For most players, this isn't that big of a deal. But if the
player were a Drow and fighting with their dark elven weaponry, there's
a non-zero chance that their weapons will crumble 'in the light', and
this was reported a few times as having happened.

Turns out the game would remember the last lit/unlit state the player
was standing in before being engulfed. Current fix is to add a bit to
spot_is_dark() so when engulfed, the player is always considered being
in the dark.

A couple things should be accounted for, such as: does the engulfer emit
light (maybe not have a case for a gold dragon, but definitely for
something like a fire vortex/elemental). And magic lamps are not snuffed
out when engulfed, so technically the player could have light while
engulfed still from a blessed/uncursed magic lamp, or while wearing
gold dragon scales, and so on. Something for later.