New monsters: Ent, Elder Ent, Petrified Ent.
This is the first in a series of new monsters being created to flesh out
the new Druid role, its quest, and forms it may want to shapechange
into.

The Ent - from Tolkien lore, also known as a Treant in Dungeons &
Dragons. An Ent is a tree-like being, somewhat humanoid in nature, but
still very much like a tree. They are very large, move slow and have a
decent amount of AC, but are extremely strong and can hit very hard.
Ents resemble trees; if one does not pay attention, they could easily
mistake an Ent for a regular tree. As with any tree or plant, Ents are
vulnerable to fire, and they are resistant to poison and sleep attacks.

Ents do not need to eat. Their body shape and function does not allow
them to wear any armor or wield weapons, but amulets and rings are still
allowed, and they will pick up and utilize useful items. They are also
large enough and strong enough to pickup and toss boulders (remember the
Battle of Isengard? Heh).

Ents can grow up into Elder Ents, which are larger and more powerful.
The strongest Ent of them all is the Petrified Ent (this one is my own
creation, and doesn't exist anywhere else as far as I can tell).
Petrified Ents are Elder Ents that became passive/meditated for so long,
they became petrified. These Ents are no longer made of wood, but of
stone, and are resistant to stoning. Elder Ents will not grow up into
Petrified Ents as how they come to be is not exactly natural.

When an Ent or Elder Ent is killed, they leave behind a dead tree as a
corpse. A Petrified Ent will leave behind a statue of itself upon dying,
just like stone golems do. Stoning an Ent or an Elder Ent will turn it
into a Petrified Ent; casting 'stone to flesh' at a Petrified Ent will
turn it into a regular Ent.

Ents are always peaceful towards Druids, Rangers, and any role that is
of the elven race (except Infidels).

There's probably more I can do here, such as 'if an Ent sees the player
harming a tree in any way, they become hostile'. Anywho. Enjoy the new
monsters.