New monster: Satyr.
This commit adds the Satyr to the game. Also known as fauns, Satyrs are
male humanoids closely releated to other fey creatures, such as nymphs.
Appearance-wise, they appear as half-man, half-goat, having fur covering
the lower parts of their bodies, and cloven hooves. They also have a
pair of horns atop their head. Satyrs are always male; as far as anyone
can surmise, female satyrs just don't occur naturally, which seems a bit
odd.

The Satyr for NetHack first appeared in the biodiversity patch, many
years ago. That patch is featured in SlashTHEM, but the implementation
to me seemed a bit convoluted. So this is my stab at it, hopefully I do
it some justice.

Sayrs appear as an orange n, and are considered part of the nymph family
in most regards. They do not have a nymphs ability to steal or seduce,
or have their massive carrying cap, nor do they teleport. Instead, they
have two physical attacks: weapon, and a head butt using their horns.
They have a chance to spawn with a small variety of elven weaponry.

One distinguishing feature of Satyrs is their ability to affect their
enemies with music. They will spawn with a new object, the pan flute,
80% of the time. Periodically, Satyrs will play their flute, which can
cause a handful of different outcomes - sleep, stun, or confuse. The
duration of each effect only lasts a few turns, and confuion/stun will
not stack with its already existing effect. If the player is wearing a
toque, is deaf, or is polymorphed into any monster of the nymph class,
they will be unaffected.

The pan flute does not generate randomly, and will only even appear with
a spawned satyr, or maybe found in a bones pile. The player can't use a
pan flute unless they themselves polymorph into satyr. In doing so, they
can then use the pan flute to either put monsters to sleep, or charm
them.

Could maybe do a bit more with this new addition, but this is good for
the time being.