Fix: uncontrolled levelporting below the Valley of the Dead while Cerberus was still alive.
This was discovered very recently during a round robin game of EvilHack
(they ascended, yay). Uncontrolled levelporting below the Valley of the
Dead was not accounted for when shutting off levelporting there while
Cerberus was still alive. I had the block of code handling that in the
wrong spot. To fix, move it up the chain before controlled levelporting
is handled.

Also did a bit of code formatting and added a couple wizard mode
overrides.