Skeletons don't have a 'thick hide' exactly.
Skeletons have the M1_THICK_HIDE flag, meaning their body affords some
natural protection from physical attacks. Just recently watched my pet
attack a skeleton, and the skeletons 'thick hide' blocked the attack.
That obviously is not the correct feedback. So fixed that, and also
noticed in trap.c that the various types of hides and such weren't
accounted for when deflecting certain traps, so that's fixed as well.