New monster: Assassin Vine.
Another plant-based monster, the Assassin Vine. This plant-based monster
was first introduced in ad&d third edition, and has been somewhat ported
over to EvilHack. Its attacks and how it behaves is kind of a mix
between dNetHack's Razorvine, and UnNetHack's Devil's Snare, and I've
borrowed a bit of code from both.

Assassin vines appear as { when using the default/plain symset, and as £
when using DEC/IBMgraphics/curses symset, and are bright green in color.
They are immobile, but can still actively attack if the player gets
right next to one (melee range). These vines have two attacks: first is
a physical touch attack (1d4), the second is a grab attack that can
ensnare and then crush the player (3d6). Assassin vines are immune to
sleep, poison, sickness, and electric-based attacks.

The one big thing that sets this monster apart from almost every other
monster, is its ability to 'grow' throughout the dungeon. The main
individual vine will slowly start to spread into any available tile
space that is determined to be good (meaning it won't grow into walls,
etc). The vine clusters that spawn off of the main vine will have
one-half to one-third of the hit points of the main vine, but if left
alone, will regenerate hit points up to its max. The vines that grow
from the main vine will not spawn their own vines - only the main
starting vine grows. To prevent a cluster of assassin vines from taking
over its area, the main vine either needs to be killed, or at the very
least have its hit points dropped to two-thirds of max or lower (it will
start growing again if allowed to regenerate its hit points above the
two-thirds threshold).

Assassin vines aren't too dangerous by themselves - they are level 10
and small, so the amount of hit points the vine has isn't too great, and
their AC makes them fairly easy to hit. But if caught in a group of
them, and/or having to deal with another monster at the same time while
entangled in the vines, it can set the stage for a dangerous scenario.