New level and new special room (forest).
This addition will give players other than a Druid a chance to
experience the new tree and plant-like creatures in their natural
habitat.

Both the new level and the new room type are referred to as a 'forest',
which are filled with grass, and some dead trees and shallow water. The
monsters that can inhabit this level/room are mostly blights, but the
occasional assassin vine or ent could appear as well.

The new level can randomly spawn between levels 12 and 22 (inclusive).
There is a boss monster on this level - a Gulthias Tree. This tree is
unique, and is the reason why all other blights exist. It has three
attacks (two physical (claw) and one grasping attack that can immobilize
and crush its target) and is stationary. Periodically it will spawn
either a twig blight or a needle blight. To prevent these blights from
taking over the forest, the Gulthias Tree must be killed.

Also in this commit, true ents (not just all S_ENT) and blights will
grudge each other.