Complete the Druid quest levels.
This commit finally completes work on the Druid quest levels. The theme
is a forest that is being corrupted by the quest nemesis (Baba Yaga) and
the transition from live to dead trees becomes more pronounced the
deeper you go. Planty of dangers await the Druid as they attempt to
complete their quest, with plenty of the new monsters recently
added making appearances here.

It's very possible for some tweaks and changes to happen with these
levels. I'm not entirely happy with the middle 'local' level and how
Baba Yaga's hut is depicted. Per lore, it's a mobile hut on two giant
chicken legs... can't do that here. We'll see.

Also in this commit: the Druid quest leader and guardians are properly
kitted out with appropriate gear. In testing this, I discovered that
monsters can't wield a two-handed weapon while wearing bracers, so that
will be addressed in a future commit.

Last bit - Druids can always move diagonally between trees (dead or
alive).