Satyrs enjoy booze
Satyrs will seek out booze and catch any that is thrown at them. Drinking non-cursed potions of booze will pacify them. Cursed potions will send them into a berserk rage (more appropriate to the maenads than satyrs, really, but I feel like berserking is an underutilized monster status and jumped at the chance to incorporate it). Either way, they will be confused, giving you a reason to give them some space and/or time to escape before they come after you.

Hopefully this doesn't make it *too* easy to avoid satyrs. The pacification can always be downgraded to only have a chance of it happening, or perhaps the catching mechanic can be removed so it's harder to get the booze to them if it turns out to be too easy.