New spells: barkskin and stoneskin.
This commit adds the first two new spells for Druid role and evocation
line of spells - barkskin and stoneskin. Both spells are
non-directional, and can only be cast on the player (for now at least).

Barkskin is a common ad&d spell for Druids, and in that setting, it is a
level two spell. But for EvilHack, it is a level one spell, and has some
things in common with the first level clerical spell 'protection'.
Here's the breakdown:

* Barkskin is a level one evocation spell
* At basic skill in evocation spells, it provides an extra 5 AC. At
skilled, it's an additional 3 points of extra AC (8 total), and at
expert, and additional 2 points of extra AC on top of that (10 total).
Cast unskilled/restricted, it only provides one point of AC
* The amount of turns the spell lasts is the same as the matter spell
'reflection' (skill based), and the timeout can be extended by
repeatedly casting the spell
* Unlike the clericial spell 'protection', the player does not gain
extra AC by recasting the spell repeatedly. However, the AC gained does
not gradually decrease either
* Having this spell active is akin to being thick-skinned
* Various types of attacks have their damage reduced (or in same edge
cases, increased) while barkskin is active

Stoneskin is also found in ad&d, but was listed as a magic-users
(wizard) spell. The details:

* Stoneskin is a level four evocation spell
* At basic skill in evocation spells, it provides an extra 8 AC. At
skilled, it's an additional 4 points of extra AC (12 total), and at
expert, and additional 3 points of extra AC on top of that (15 total).
Cast unskilled/restricted, it only provides 3 points of AC
* The amount of turns the spell lasts is the same as the matter spell
'reflection' (skill based), and the timeout can be extended by
repeatedly casting the spell
* Unlike the clericial spell 'protection', the player does not gain
extra AC by recasting the spell repeatedly. However, the AC gained does
not gradually decrease either
* Having this spell active is akin to being thick-skinned
* Various types of attacks have their damage reduced while stoneskin is
active
* The player is stone resistant while the stoneskin spell is active

A select few spellcasting monsters can cast both of these spells on
themselves also (player monster druids, and the druid quest leader and
her guardians).

Both spells when active protect the player (and monsters) from materials
they hate.

The way this was setup, it would be possible to create objects that
grant these protections as well. But for now we'll just stick to the
spells.