New spell: entangle.
New evocation/druid spell - entangle. This is a level one evocation
spell (barksin has been bumped up to level two) and is currently a
druids guaranteed starting spell, and is also considered its emergency
spell.

When cast at a monster, it will cause surrounding plant life (this is
all assumed of course, we're in a dungeon for the most part) to entangle
and hold the target for a brief number of turns. The affected target can
still do things, such as melee attack the player if they are adjacent,
but they are otherwise immobile for the duration of the spell.

Spell duration depends on a couple factors - the evocation spell skill
level of the caster, and whether or not the target of the spell is on a
grass tile. For example, at basic skill, the entanglement lasts for 3-6
turns, or 3-8 if they are on a grass tile. The upper range goes up
slightly with skill. At unskilled/restricted, the amount of turns is low
(1-4 turns, or 1-6 on grass).

If the caster is skilled or greater in evocation spells, there's a 50%
chance of temporarily slowing the target while entangled.

An entangled monsters behavior will match that of a trapped monster in
many cases.

There's room for improvement here. Currently this isn't a monster spell
that can be used against the player, but the player could cast this at
themselves, and there's some handling for that. Something else that
would be neat, is if surrounding terrain (read: trees) could be
accounted for. Could also take into account very strong and/or
huge/gigantic monsters.

Will be interesting to see how this spell is used in a real game. Great
spell to use to aid in escaping, or just to buy some time.