New spell: change metal to wood.
Another new Druid spell - change metal to wood. This is a fifth level
evocation spell. The way it works: the caster either wields want they
want to change, or they can also target objects on the floor, or even
other monsters. If successful, the object will transmute into wood.

There are some restrictions to this spell - first, the targeted object
has to be made of metal. It also has to be possible for it to be made of
wood in the first place (example: this won't work on a shield of
reflection, as those can never spawn being made of wood). Lastly -
artifacts, invocation objects, and the Amulet of Yendor are off-limits.

Targeting an iron golem will change it into a wood golem.

The player will need to be careful using this spell in a shop with
unpaid items, as changing the material will incur a cost.

Why would a Druid want to use this spell? The main reason is that (and
this isn't implemented yet, but will be soon) if a Druid wears any kind
of metallic armor, they can't cast spells or use their special
abilities. There's also the benefit of reducing the weight of various
objects that perform just fine regardless of their material, as wood
tends to be lighter than most metals.