New object: staff of evocation.
With the addition of the new spell skill/class (evocation), it warrants
an additional magical staff for that particular spell type - the staff
of evocation.

The staff of evocation appears as a 'gnarled staff' and is made of wood.
It grants a +1 bonus to-hit, and when wielded, gives a 50% bonus to
casting success for evocation spells. The Staff of the Archmagi also
acquires the same bonus.