Druids and bonuses/penalties vs certain materials.
If a Druid wears any sort of armor that is made of any type of metal,
they will be unable to cast spells. This will also extend to
future special abilities (read: polyform) that have not been created
yet.

Druids gain a single point AC bonus per piece of armor worn that is made
of wood.

Druids receive a +2 bonus to damage when attacking with any weapon made
of wood.