Druids, elves are penalized for killing trees.
If an elven player kills a live tree, they will take a slight hit to
alignment and luck. If a Druid does the same thing, it's super bad -
major hit to alignment and luck, and their deity becomes 'very' angry.

These penalties do not stack if the player is an elven Druid.