Druids that #shapechange retain worn armor, other changes.
One of the things that makes a Druid using their shapechange ability in
ad&d (it's called 'wildshape' there) fun is that they have the ability for
their worn armor and/or weapons to remain and magically 'merge' with the
Druids form. How this works, if even allowed at all, is at the DM's
discretion. Every time I played a Druid in ad&d, the DM always allowed
the armor to merge - this was a big part of the appeal, otherwise why
bother doing it? Going to apply the same logic here.

When a Druid uses #shapechange, these things will now occur:
* any worn armor will 'merge' with their form and will stay worn/in use
if their new shape wouldn't otherwise allow it
* if you remove said armor while shapechange is active, you won't be
able to put that armor back on until you revert to normal form
* weapons will still be unusable if in a form that can't wield them, but
they will stay in inventory and not be dropped instead
* carry cap (amount of weight) has been adjusted so that the player can
still realistically maintain what they have in inventory regardless of
the form they chose - the magic imparted by their deity gives them the
same carry cap as if they were wearing gauntlets of power

Still more to do here, but the feasability and appeal of wanting to
use #shapechange just went way up.