Rework Druid shapechange form defines.
Doing this rework/restructuring because there are certain behaviors that
shouldn't rely on the u.ushapechange timer. This addresses an issue I
noticed where if the player Druid shapechanged, then reverted, but was
then polymorphed into a monster by normal means and u.ushapechange was a
non-zero value, they would still enjoy all the aspects of armor merging,
carry cap, etc. Not ideal.

Co-authored-by: saltwaterterrapin <saltwaterterrapin@users.noreply.github.com>