Fix: sewage and shallow water tiles were lit in makeriver().
Way back near the beginning of EvilHack development, rivers were added
to the gnomish mines (from GruntHack). Unlike that variant, I made sure
to make the pools and moats generated as lit so the player wouldn't
blindly stumble into them if a mines level were dark. Later on when
sewage and shallow water were introduced and added to the makeriver()
routine, I made them lit as well.

And now we have Drow as a playable race, and if you've been following
EvilHack development and/or playing it, you know that it's not a good
idea for Drow to be standing in a lit spot. This commit changes things
slightly - sewage and shallow water are no longer lit in that routine.
There's really no good reason for it, as the player can safely stand on
those spots 999,999 out of 1,000,000 times. And it was a player playing
as Drow that lost their main dark elven weapon on the sewage level
(mines' end) even though that level is dark, because all of the sewage
tiles are lit.

Something else I just thought of while typing out this commit. Since
paranoid swim is a thing now, pools/moats don't really need to be lit
in the mines either. I'll think on that one for a bit.